wazieat_admin/requirements.txt
asgiref==3.4.1
attrs==21.4.0
python-Levenshtein-wheels==0.13.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
django-rest-swagger==2.2.0
django-storages==1.12.3
django-tenants==3.4.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
drf-writable-nested==0.6.3
drf-yasg==1.20.0
geographiclib==1.52
geopy==2.2.0
google==3.0.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyrsistent==0.18.1
python-gettext==4.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.8.0
pyshorteners==1.0.1
